Added Circle class and touch code for multi touch and circle drawing code.
For now the code to determine the center of the circle and the radius are just stubbed in, but I will finish that up tomorrow.  Also the circle stays red and never changes to black after the touch event ends, I suspect this is because the [event allTouches] isn't returning nil, need to see why.